add edit links to collection details page
