+++ app/components/collections/detail_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th scope="col" class="table-title col-3">Details</th>
+      <th scope="col" class="table-title col-3">Details<%= edit_link('#', "Edit details") %></th>
       <td scope="col"></td>
     </tr>
   </thead>
