+++ app/components/collections/header_component.html.erb
@@ -1,5 +1,6 @@
 <header class="title">
   <span class="header-text"><%= name %></span>
+  <%= edit_link('#', "Edit #{name}") %>
   <% if depositing? %>
     <span data-state-updates-target="state" class="state">
       Depositing <%= spinner %>
