+++ app/components/collections/links_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th class="table-title col-3">Links to related information</th>
+      <th class="table-title col-3">Links to related information<%= edit_link('links', "Edit links to related information") %></th>
       <td></td>
     </tr>
   </thead>
