+++ app/components/collections/participants_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th scope="col" class="table-title col-3">Collection participants</th>
+      <th scope="col" class="table-title col-3">Collection participants<%= edit_link('participants', "Edit collection participants") %></th>
       <td scope="col"></td>
     </tr>
   </thead>
