+++ app/components/collections/participants_component.rb
@@ -3,15 +3,7 @@
 
 module Collections
   # Renders the participant section of the collection (show page)
-  class ParticipantsComponent < ApplicationComponent
-    sig { params(collection: Collection).void }
-    def initialize(collection:)
-      @collection = collection
-    end
-
-    sig { returns(Collection) }
-    attr_reader :collection
-
+  class ParticipantsComponent < Collections::ShowComponent
     sig { returns(T.nilable(String)) }
     def depositors
       collection.depositors.map(&:sunetid).join(', ')
