+++ app/components/collections/release_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th scope="col" class="table-title col-3">Release and visibility</th>
+      <th scope="col" class="table-title col-3">Release and visibility<%= edit_link('release_details', "Edit release options") %></th>
       <td scope="col"></td>
     </tr>
   </thead>
