+++ app/components/collections/release_component.rb
@@ -3,15 +3,7 @@
 
 module Collections
   # Renders the release section of the collection (show page)
-  class ReleaseComponent < ApplicationComponent
-    sig { params(collection: Collection).void }
-    def initialize(collection:)
-      @collection = collection
-    end
-
-    sig { returns(Collection) }
-    attr_reader :collection
-
+  class ReleaseComponent < Collections::ShowComponent
     delegate :release_option, :access, to: :collection
   end
 end
