+++ app/components/collections/terms_of_use_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th scope="col" class="table-title col-3">Terms of use and licenses</th>
+      <th scope="col" class="table-title col-3">Terms of use and licenses<%= edit_link('terms', "Edit terms of use and license") %></th>
       <td scope="col"></td>
     </tr>
   </thead>
