+++ app/components/collections/workflow_review_component.html.erb
@@ -1,7 +1,7 @@
 <table class="table table-sm mb-5">
   <thead class="table-light">
     <tr>
-      <th scope="col" class="table-title col-3">Review workflow</th>
+      <th scope="col" class="table-title col-3">Review workflow<%= edit_link('review', "Edit review workflow") %></th>
       <td scope="col"></td>
     </tr>
   </thead>
